requirements.txt
streamlit>=1.30.0
torch>=2.0.0
opencv-python-headless
numpy
Pillow
pyyaml
matplotlib
tqdm
seaborn
scipy
requests
